add-go-fuse-to-inject-filesystem-error
